follow module

+++ src/user/application/services/delete-user-address.service.ts
@@ -1,4 +1,4 @@
-import { Inject, Injectable } from "@nestjs/common";
+import { Inject, Injectable, UnauthorizedException } from "@nestjs/common";
 import { DeleteUserAddressUseCase } from "../ports/in/use-cases/delete-user-address.use-case";
 import { UserAddressRepository, UserAddressRepositoryProvider } from "../ports/out/user-address.repository";
 import { UserRepository, UserRepositoryProvider } from "../ports/out/user.repository";
@@ -7,6 +7,7 @@ import { User } from "src/user/domain/user.model";
 import { UserNotFoundError } from "./errors/user-not-found.error";
 import { UserAddress } from "src/user/domain/user-address.model";
 import { UserAddressNotFoundError } from "./errors/user-address-not-found.error";
+import { GetAuthenticatedUserUseCase, GetAuthenticatedUserUseCaseProvider } from "src/auth/application/ports/in/use-cases/get-authenticated-user.use-case";
 
 @Injectable()
 export class DeleteUserAddressService implements DeleteUserAddressUseCase 
@@ -15,7 +16,9 @@ export class DeleteUserAddressService implements DeleteUserAddressUseCase
         @Inject(UserAddressRepositoryProvider) 
         protected userAddressRepository: UserAddressRepository,
         @Inject(UserRepositoryProvider) 
-        protected userRepository: UserRepository
+        protected userRepository: UserRepository,
+        @Inject(GetAuthenticatedUserUseCaseProvider)
+        protected getAuthenticatedUser: GetAuthenticatedUserUseCase
     ) 
     {}
 
@@ -23,20 +26,22 @@ export class DeleteUserAddressService implements DeleteUserAddressUseCase
     {
         const user: User = await this.userRepository.findById(command.userId);
         
-        if(user == null) {
-            throw new UserNotFoundError(`User ${command.userId} not found.`);
+        if(user === null || user === undefined || user.isDeleted()) {
+            throw new UserNotFoundError(`User ${command.userId} not found`);
+        }
+
+        if(user.id() !== this.getAuthenticatedUser.execute(null)) {
+            throw new UnauthorizedException(`Unauthorized access`);
         }
 
         const userAddress: UserAddress = await this.userAddressRepository.findById(command.id);
 
-        if(userAddress == null || userAddress.isDeleted() || userAddress.user().id() != user.id()) {
-            throw new UserAddressNotFoundError(`User address ${command.id} not found.`);
+        if(userAddress === null || userAddress === undefined || userAddress.isDeleted() || userAddress.user().id() !== user.id()) {
+            throw new UserAddressNotFoundError(`User address ${command.id} not found`);
         }
 
         userAddress.delete();
 
         this.userAddressRepository.update(userAddress);
-    
-        return null;
     }
 }
